requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
SQLAlchemy==2.0.36
alembic==1.13.3
python-dotenv==1.0.1
aiofiles==24.1.0
Pillow==11.0.0
loguru==0.7.2
redis==5.2.0
aioredis==2.0.1
apscheduler==3.10.4
babel==2.16.0
